+++ apiserver/service/src/main/java/com/zipline/service/region/RegionService.java
@@ -1,8 +1,10 @@
 package com.zipline.service.region;
 
 import com.zipline.global.response.ApiResponse;
-import com.zipline.service.region.dto.RegionResponseDTO;
+import com.zipline.service.region.dto.FlatRegionDTO;
+
+import java.util.List;
 
 public interface RegionService {
-    public ApiResponse<RegionResponseDTO> getChildrenRegions(Long cortaNo);
+    public ApiResponse<List<FlatRegionDTO>> getChildrenRegions(Long cortaNo);
 }
